src/main/resources/assets/examplemod/textures/gui/requirements.txt
Pillow==5.4.1
numpy==1.8.0rc1